basics of new page animations
very rudimentary as of right now, but functional. tried to make it as easy to play around with as possible - check out the comments in OneConfigGui.java in the openPage method

still left in some of the basic code for the horizontal transition, and added isX for animations to make an animation that works for both x and y directions. could be removed, added for convenience atm